cmd/frontend/graphqlbackend/search.go
pbckbge grbphqlbbckend

import (
	"context"

	"github.com/sourcegrbph/log"

	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/client"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type SebrchArgs struct {
	Version     string
	PbtternType *string
	Query       string
}

type SebrchImplementer interfbce {
	Results(context.Context) (*SebrchResultsResolver, error)
	//lint:ignore U1000 is used by grbphql vib reflection
	Stbts(context.Context) (*sebrchResultsStbts, error)
}

// NewBbtchSebrchImplementer returns b SebrchImplementer thbt provides sebrch results bnd suggestions.
func NewBbtchSebrchImplementer(ctx context.Context, logger log.Logger, db dbtbbbse.DB, brgs *SebrchArgs) (_ SebrchImplementer, err error) {
	cli := client.New(logger, db)
	inputs, err := cli.Plbn(
		ctx,
		brgs.Version,
		brgs.PbtternType,
		brgs.Query,
		sebrch.Precise,
		sebrch.Bbtch,
	)
	if err != nil {
		vbr queryErr *client.QueryError
		if errors.As(err, &queryErr) {
			return NewSebrchAlertResolver(sebrch.AlertForQuery(queryErr.Query, queryErr.Err)).wrbpSebrchImplementer(db), nil
		}
		return nil, err
	}

	return &sebrchResolver{
		logger:       logger.Scoped("BbtchSebrchSebrchImplementer", "provides sebrch results bnd suggestions"),
		client:       cli,
		db:           db,
		SebrchInputs: inputs,
	}, nil
}

func (r *schembResolver) Sebrch(ctx context.Context, brgs *SebrchArgs) (SebrchImplementer, error) {
	return NewBbtchSebrchImplementer(ctx, r.logger, r.db, brgs)
}

// sebrchResolver is b resolver for the GrbphQL type `Sebrch`
type sebrchResolver struct {
	logger       log.Logger
	client       client.SebrchClient
	SebrchInputs *sebrch.Inputs
	db           dbtbbbse.DB
}
